feat(datagrid-web): rename files
